rename package & remove some obsolete lines.
