+++ bfe_wasmplugin/plugin.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package bfe_wasmplug
+package bfe_wasmplugin
 
 import (
 	"crypto/md5"
@@ -26,7 +26,7 @@ import (
 	"sync/atomic"
 
 	"github.com/baidu/go-lib/log"
-	wasmABI "github.com/bfenetworks/bfe/bfe_wasmplug/abi"
+	wasmABI "github.com/bfenetworks/bfe/bfe_wasmplugin/abi"
 	"github.com/bfenetworks/proxy-wasm-go-host/proxywasm/common"
 	v1Host "github.com/bfenetworks/proxy-wasm-go-host/proxywasm/v1"
 )
